fix: desktop click away bug
