+++ src/common/components/elements/UPopper.tsx
@@ -1,9 +1,8 @@
 import { styled } from '@/common/lib/mui/theme'
 import Popper from '@mui/material/Popper'
-import ClickAwayListener from '@mui/material/ClickAwayListener'
 import type React from 'react'
-import { useCallback, useMemo } from 'react'
 import HeaderPopperOverlay from '@/common/components/elements/Header/HeaderPopperOverlay'
+import UClickAwayListener from '@/common/components/elements/UClickAwayListener'
 
 const StyledPopper = styled(Popper)(() => ({
   display: 'flex',
@@ -29,38 +28,12 @@ const UPopper = ({
   clickAwayClassNameWhiteList,
   onClose,
 }: UPopperProps) => {
-  const clickAwayClassNameWhiteListSet = useMemo(
-    () => new Set(clickAwayClassNameWhiteList),
-    [clickAwayClassNameWhiteList]
-  )
-
-  const handleClickAway = useCallback(
-    (event: MouseEvent | TouchEvent) => {
-      /**
-       * 從 event.target 往上找，直到找到 clickAwayClassNameWhiteList 中的 class name
-       * 如果找到，則不觸發 onClickAway
-       */
-      let parent = event.target as HTMLElement
-      while (parent) {
-        if (
-          Array.from(parent.classList).some((className) =>
-            clickAwayClassNameWhiteListSet.has(className)
-          )
-        ) {
-          return
-        }
-        parent = parent.parentElement as HTMLElement
-      }
-      if (!parent) {
-        onClose?.()
-      }
-    },
-    [clickAwayClassNameWhiteListSet, onClose]
-  )
-
   return (
     <>
-      <ClickAwayListener onClickAway={handleClickAway}>
+      <UClickAwayListener
+        clickAwayClassNameWhiteList={clickAwayClassNameWhiteList}
+        onClickAway={onClose}
+      >
         <StyledPopper
           className={className}
           anchorEl={anchorEl}
@@ -69,7 +42,7 @@ const UPopper = ({
         >
           {children}
         </StyledPopper>
-      </ClickAwayListener>
+      </UClickAwayListener>
       <HeaderPopperOverlay open />
     </>
   )
